Creating facade for objectsByKey
