+++ jasmine.json
@@ -1,6 +1,6 @@
 {
   "spec_dir": ".",
   "spec_files": [
-    "./spec/**/*.spec.jsx"
+    "./spec/**/*.spec.js"
   ]
 }
