+++ package.json
@@ -1,10 +1,9 @@
 {
   "name": "react-entity",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "Create entities base on React propTypes",
   "main": "lib/ReactEntity.js",
   "scripts": {
-    "postinstall": "babel ./src -d lib",
     "test": "node spec/jasmine-runner",
     "coverage": "babel-istanbul cover 'spec/jasmine-runner.js' -x '**/spec/**' && open ./coverage/lcov-report/index.html"
   },
